add more packet defines

+++ TrProtocol/Packets/InstancedItem.cs
@@ -5,7 +5,7 @@ namespace TrProtocol.Packets
     public class InstancedItem : Packet, IItemSlot
     {
         public override MessageID Type => MessageID.InstancedItem;
-        public byte ItemSlot { get; set; }
+        public short ItemSlot { get; set; }
         public Vector2 Position { get; set; }
         public Vector2 Velocity { get; set; }
         public short Stack { get; set; }
