+++ TrProtocol/Packets/ItemOwner.cs
@@ -3,7 +3,7 @@
     public class ItemOwner : Packet, IItemSlot, IOtherPlayerSlot
     {
         public override MessageID Type => MessageID.ItemOwner;
-        public byte ItemSlot { get; set; }
+        public short ItemSlot { get; set; }
         public byte OtherPlayerSlot { get; set; }
     }
 }
