+++ TrProtocol/Packets/SyncItem.cs
@@ -5,7 +5,7 @@ namespace TrProtocol.Packets
     public class SyncItem : Packet, IItemSlot
     {
         public override MessageID Type => MessageID.SyncItem;
-        public byte ItemSlot { get; set; }
+        public short ItemSlot { get; set; }
         public Vector2 Position { get; set; }
         public Vector2 Velocity { get; set; }
         public short Stack { get; set; }
